Backendd/interview/requirements.txt
flask>=2.0.0
flask-cors>=3.0.0
google-cloud-speech>=2.0.0
google-cloud-storage>=2.0.0
openai>=1.0.0
python-dotenv>=0.19.0 